// ryan_testing/array_gRPC_testing/Ingress_Node/proto/contouring.ts
export interface ContouringOutput {
  value: string;
}

export interface ContouringEmpty {
  data: number[];
  width: number;
  height: number;
  offset: number;
  scale: number;
  index: number;
}

function createBaseContouringOutput(): ContouringOutput {
  return { value: "" };
}

export const ContouringOutput = {
  encode(message: ContouringOutput, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.value !== "") {
      writer.uint32(10).string(message.value);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ContouringOutput {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseContouringOutput();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.value = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ContouringOutput {
    return { value: isSet(object.value) ? globalThis.String(object.value) : "" };
  },

  toJSON(message: ContouringOutput): unknown {
    const obj: any = {};
    if (message.value !== "") {
      obj.value = message.value;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ContouringOutput>, I>>(base?: I): ContouringOutput {
    return ContouringOutput.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ContouringOutput>, I>>(object: I): ContouringOutput {
    const message = createBaseContouringOutput();
    message.value = object.value ?? "";
    return message;
  },
};

function createBaseContouringEmpty(): ContouringEmpty {
  return { data: [], width: 0, height: 0, offset: 0, scale: 0, index: 0 };
}

export const ContouringEmpty = {
  encode(message: ContouringEmpty, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    writer.uint32(10).fork();
    for (const v of message.data) {
      writer.float(v);
    }
    writer.ldelim();
    if (message.width !== 0) {
      writer.uint32(21).float(message.width);
    }
    if (message.height !== 0) {
      writer.uint32(29).float(message.height);
    }
    if (message.offset !== 0) {
      writer.uint32(37).float(message.offset);
    }
    if (message.scale !== 0) {
      writer.uint32(45).float(message.scale);
    }
    if (message.index !== 0) {
      writer.uint32(53).float(message.index);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ContouringEmpty {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseContouringEmpty();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag === 13) {
            message.data.push(reader.float());

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.data.push(reader.float());
            }

            continue;
          }

          break;
        case 2:
          if (tag !== 21) {
            break;
          }

          message.width = reader.float();
          continue;
        case 3:
          if (tag !== 29) {
            break;
          }

          message.height = reader.float();
          continue;
        case 4:
          if (tag !== 37) {
            break;
          }

          message.offset = reader.float();
          continue;
        case 5:
          if (tag !== 45) {
            break;
          }

          message.scale = reader.float();
          continue;
        case 6:
          if (tag !== 53) {
            break;
          }

          message.index = reader.float();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ContouringEmpty {
    return {
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => globalThis.Number(e)) : [],
      width: isSet(object.width) ? globalThis.Number(object.width) : 0,
      height: isSet(object.height) ? globalThis.Number(object.height) : 0,
      offset: isSet(object.offset) ? globalThis.Number(object.offset) : 0,
      scale: isSet(object.scale) ? globalThis.Number(object.scale) : 0,
      index: isSet(object.index) ? globalThis.Number(object.index) : 0,
    };
  },

  toJSON(message: ContouringEmpty): unknown {
    const obj: any = {};
    if (message.data?.length) {
      obj.data = message.data;
    }
    if (message.width !== 0) {
      obj.width = message.width;
    }
    if (message.height !== 0) {
      obj.height = message.height;
    }
    if (message.offset !== 0) {
      obj.offset = message.offset;
    }
    if (message.scale !== 0) {
      obj.scale = message.scale;
    }
    if (message.index !== 0) {
      obj.index = message.index;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ContouringEmpty>, I>>(base?: I): ContouringEmpty {
    return ContouringEmpty.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ContouringEmpty>, I>>(object: I): ContouringEmpty {
    const message = createBaseContouringEmpty();
    message.data = object.data?.map((e) => e) || [];
    message.width = object.width ?? 0;
    message.height = object.height ?? 0;
    message.offset = object.offset ?? 0;
    message.scale = object.scale ?? 0;
    message.index = object.index ?? 0;
    return message;
  },
};

export type ContourServicesService = typeof ContourServicesService;
export const ContourServicesService = {
  computeContour: {
    path: "/ContourServices/computeContour",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: ContouringEmpty) => Buffer.from(ContouringEmpty.encode(value).finish()),
    requestDeserialize: (value: Buffer) => ContouringEmpty.decode(value),
    responseSerialize: (value: ContouringOutput) => Buffer.from(ContouringOutput.encode(value).finish()),
    responseDeserialize: (value: Buffer) => ContouringOutput.decode(value),
  },
} as const;

export interface ContourServicesServer extends UntypedServiceImplementation {
  computeContour: handleUnaryCall<ContouringEmpty, ContouringOutput>;
}

export interface ContourServicesClient extends Client {
  computeContour(
    request: ContouringEmpty,
    callback: (error: ServiceError | null, response: ContouringOutput) => void,
  ): ClientUnaryCall;
  computeContour(
    request: ContouringEmpty,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: ContouringOutput) => void,
  ): ClientUnaryCall;
  computeContour(
    request: ContouringEmpty,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: ContouringOutput) => void,
  ): ClientUnaryCall;
}

export const ContourServicesClient = makeGenericClientConstructor(
  ContourServicesService,
  "ContourServices",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): ContourServicesClient;
  service: typeof ContourServicesService;
  serviceName: string;
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };